+++ index.ts
@@ -1,13 +1,14 @@
 import express, { Express } from "express";
-import * as bodyParser from "body-parser";
 
 import authRoutes from "./routes/auth.route";
 import config from "./config/app.config";
 import dbConnect from "./utils/databaseConnect";
+import cors from "./middlewares/cors.middleware";
 
 const app: Express = express();
 
-app.use(bodyParser.json());
+app.use(express.json());
+app.use(cors);
 
 dbConnect(config.db.MONGODB_URI);