+++ interfaces/Token.interface.ts
@@ -1,0 +1,6 @@
+interface Token {
+	username: string;
+	refreshToken: string;
+}
+
+export default Token;