+++ middlewares/cors.middleware.ts
@@ -1,0 +1,16 @@
+import cors from "cors";
+
+const options: cors.CorsOptions = {
+  allowedHeaders: [
+    "Origin",
+    "X-Requested-With",
+    "Content-Type",
+    "Accept",
+    "X-Access-Token",
+  ],
+  credentials: true,
+  origin: "API_URL",
+  preflightContinue: false,
+};
+   
+export default cors(options);