+++ routes/post.route.ts
@@ -1,3 +1,12 @@
-import router from "../utils/router";
+import { Request, Response } from 'express';
 
-router.get("/", (res, req) => {});
+import router from '../utils/router';
+import authorize from '../middlewares/authorize.middleware';
+
+router.get('/', (_req: Request, _res: Response) => {});
+
+router.post('/create', authorize, (req: Request, res: Response) => {
+	res.json({ req: req.body });
+});
+
+export default router;